# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Weekly Commitments API
  description: |
    API for the Weekly Commitments module – plan creation, lifecycle management,
    reconciliation, manager review, RCDO integration, and AI-assisted suggestions.
  version: 1.0.0
  contact:
    name: ST6 Engineering
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
...
  /commits/{commitId}/check-in:
    post:
      operationId: addCheckIn
      summary: Append a daily check-in to a commit
      description: |
        Appends a structured progress micro-update to the given commit.
        Check-in history is append-only — entries are never modified or deleted.
      tags: [Commits]
      parameters:
        - $ref: '#/components/parameters/CommitIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInRequest'
      responses:
        '201':
          description: Check-in entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /commits/{commitId}/check-ins:
    get:
      operationId: getCheckIns
      summary: Get check-in history for a commit
      description: |
        Returns the complete append-only check-in history for the given commit,
        ordered oldest-first.
      tags: [Commits]
      parameters:
        - $ref: '#/components/parameters/CommitIdPath'
      responses:
        '200':
          description: Check-in history returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /notifications/unread:
    get:
      operationId: getUnreadNotifications
      summary: Get unread in-app notifications
      description: |
        Returns unread notifications for the current user, newest first.
        MVP renders in-app banners on page load.
      tags: [Notifications]
      responses:
        '200':
          description: Unread notifications returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      summary: Mark a notification as read
      tags: [Notifications]
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notification marked as read
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      summary: Mark all notifications as read
      tags: [Notifications]
      responses:
        '204':
          description: All notifications marked as read
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /integrations/link-ticket:
    post:
      operationId: linkTicket
      summary: Link a commit to an external ticket
      description: |
        Links a weekly commit to a ticket in an external issue-tracker (Jira or Linear).
        When the commit's title or description is empty, the values are auto-populated
        from the fetched ticket data.

        Idempotent: if the link already exists, returns the existing record with 201.
      tags: [Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkTicketRequest'
      responses:
        '201':
          description: Ticket linked (or already linked — idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalTicketLink'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /commits/{commitId}/linked-tickets:
    get:
      operationId: getLinkedTickets
      summary: Get external ticket links for a commit
      description: Returns all external ticket links associated with the given weekly commit.
      tags: [Integrations, Commits]
      parameters:
        - name: commitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The weekly commit UUID
      responses:
        '200':
          description: List of linked tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedTicketsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /integrations/webhook/{provider}:
    post:
      operationId: handleIntegrationWebhook
      summary: Handle an external ticket status webhook
      description: |
        Receives a provider webhook payload and fan-outs any mapped ticket status
        change to linked internal check-in entries. This endpoint is unauthenticated
        because external providers cannot present a user session; provider-specific
        signature verification is handled separately.
      security: []
      tags: [Integrations]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
          description: External issue-tracker provider (case-insensitive; e.g. jira or linear)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Provider-specific webhook payload
      responses:
        '200':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CommitIdPath:
      name: commitId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Notification:
      type: object
      required: [id, type, payload, read, createdAt]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        payload:
          type: object
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CheckInRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ON_TRACK, AT_RISK, BLOCKED, DONE_EARLY]
          description: Current progress status of the commit
        note:
          type: string
          description: Optional free-text note (stored as empty string if omitted)
    CheckInEntry:
      type: object
      required:
        - id
        - commitId
        - status
        - note
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        commitId:
          type: string
          format: uuid
        status:
          type: string
          enum: [ON_TRACK, AT_RISK, BLOCKED, DONE_EARLY]
        note:
          type: string
        createdAt:
          type: string
          format: date-time
    CheckInHistoryResponse:
      type: object
      required:
        - commitId
        - entries
      properties:
        commitId:
          type: string
          format: uuid
        entries:
          type: array
          items:
            $ref: '#/components/schemas/CheckInEntry'
          description: Append-only history ordered oldest-first
...
    LinkTicketRequest:
      type: object
      required:
        - commitId
        - provider
        - externalTicketId
      description: |
        Request body for linking a weekly commit to an external issue-tracker ticket.
      properties:
        commitId:
          type: string
          format: uuid
          description: ID of the weekly commit to link
        provider:
          type: string
          enum: [JIRA, LINEAR]
          description: External issue-tracker provider
        externalTicketId:
          type: string
          description: Provider-specific ticket identifier (e.g. "PROJ-42" for Jira)
    ExternalTicketLink:
      type: object
      required:
        - id
        - orgId
        - commitId
        - provider
        - externalTicketId
        - createdAt
      description: A link between a weekly commit and an external issue-tracker ticket.
      properties:
        id:
          type: string
          format: uuid
          description: Link record UUID
        orgId:
          type: string
          format: uuid
          description: Organisation ID
        commitId:
          type: string
          format: uuid
          description: Weekly commit UUID
        provider:
          type: string
          enum: [JIRA, LINEAR]
          description: External issue-tracker provider
        externalTicketId:
          type: string
          description: Provider-specific ticket identifier
        externalTicketUrl:
          type: string
          nullable: true
          description: URL to the ticket in the provider's web UI
        externalStatus:
          type: string
          nullable: true
          description: Current ticket status as returned by the provider
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the most recent status sync
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the link was created
    LinkedTicketsResponse:
      type: object
      required:
        - commitId
        - links
      description: All external ticket links for a weekly commit.
      properties:
        commitId:
          type: string
          format: uuid
          description: The weekly commit UUID
        links:
          type: array
          items:
            $ref: '#/components/schemas/ExternalTicketLink'
          description: List of linked external tickets
    WebhookResponse:
      type: object
      required:
        - provider
        - checkInsCreated
      description: Result of accepting an integration webhook.
      properties:
        provider:
          type: string
          enum: [JIRA, LINEAR]
          description: External issue-tracker provider that sent the webhook
        checkInsCreated:
          type: integer
          format: int32
          minimum: 0
          description: Number of check-in entries auto-created from the webhook
...
    ApiError:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Valid JWT but insufficient access
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/ApiError'
...